Created Objects for Masav Hiuvim (Get Payment) file
Created new objects (getPaymentRecord, InstitutionGetPayment, and MasaVGetPayments) which create a masav receive payment file. The specifications for the file are found at  https://www.masav.co.il/media/1986/mifrat_hiuvim_msv.pdf

The file is almost identical to the send payment, only that the 006 code in the payment record is changed to 504. Also, there is a small re-ordering of the summary line.